Refactor out createPlanningRouteFor
